Microservice/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.21.0
pydantic==2.5.0

